commiting modified changes

+++ TimeSheetManagement/account_project_service/src/main/java/com/feuji/accountprojectservice/entity/AccountEntity.java
@@ -72,7 +72,7 @@ public class AccountEntity {
 	@Column(name="is_red")
 	private Boolean isRed;
 	@Column(name="account_status")
-	private Long accountStatus;
+	private Integer accountStatus;
 	@Column(name="comments")
 	private String comments;
 	@Column(name="is_deleted")
